+++ src/measure/imulmat.h
@@ -4,6 +4,6 @@ class IMulMat {
 public:
     virtual void init(int n, int m, int k,
                       int *la, int *lb, int *lc,
-                      double **A, double **B, double **C) = 0;
+                      float **A, float **B, float **C) = 0;
     virtual void multiply() = 0;
 };
